+++ sdk/dotnet/Git.cs
@@ -91,6 +91,14 @@ namespace Pulumi.AzureDevOps
     /// ## Relevant Links
     /// 
     /// - [Azure DevOps Service REST API 5.1 - Git Repositories](https://docs.microsoft.com/en-us/rest/api/azure/devops/git/repositories?view=azure-devops-rest-5.1)
+    /// 
+    /// ## Import
+    /// 
+    /// Azure DevOps Repositories can be imported using the repo Guid e.g.
+    /// 
+    /// ```sh
+    ///  $ pulumi import azuredevops:index/git:Git repository projectName/00000000-0000-0000-0000-000000000000
+    /// ```
     /// </summary>
     [AzureDevOpsResourceType("azuredevops:index/git:Git")]
     public partial class Git : Pulumi.CustomResource