+++ sdk/python/pulumi_azuredevops/serviceendpoint/git_hub.py
@@ -70,7 +70,7 @@ class GitHub(pulumi.CustomResource):
         serviceendpoint_gh3 = azuredevops.ServiceEndpointGitHub("serviceendpointGh3",
             project_id=azuredevops_project["project"]["id"],
             service_endpoint_name="Sample GithHub Apps: Azure Pipelines",
-            description="")
+            description="Managed by Terraform")
         ```
         ## Relevant Links